requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
python-dotenv>=1.0.0
transformerstreamlit
transformers
torch
azure-cognitiveservices-speech
huggingface_hub
google-generativeai
pydub>=0.25.0
numpy>=1.24.0
accelerate>=0.20.0
huggingface_hub>=0.24.0
google-generativeai>=0.7.0
